requirements.txt
Flask==2.2.4
Werkzeug==2.2.2
Flask-CORS==4.0.0
flasgger==0.9.7.1
flask-swagger-ui==5.21.0
gunicorn==20.0.4
python-dotenv==0.19.1
requests==2.32.3
APScheduler==3.10.4
email-validator==1.1.2
Flask-APScheduler==1.12.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.7
Flask-Breadcrumbs==0.5.1
Flask-Mail==0.10.0
Flask-JWT-Extended==4.6.0
Flask-QRcode==3.2.0
flask_menu==1.0.1
Flask-WTF==1.2.1
PyYAML==6.0.2
meshtastic==2.5.6

# Testing dependencies
pytest==8.4.2
pytest-flask==1.3.0
pytest-cov==7.0.0
requests-mock==1.12.1